-- Completed optimisation of all four (displacement/pressure-displacement    & adaptive/non-adaptive) versions of the solid element. Ought to    give this another, very final sanity check while investigating if the    pressure formulations can be made (more) symmetric while remaining    stable.
-- Fixed customise_dist.sh so that private directories are now
   included when option 1 is chosen. 

-- Modified autogen.sh so that validation for mpi-based demo drivers
   can be requested. This is done with the new flag

       --with-mpi-self-tests="mpirun -np 2"
 
   (or appropriate variants) to configure. This is most easily specified
   in the configure options file if the build is performed
   with autogen. The string must specifiy the command required
   to run an executable on two processors. 

   In future, MPI-based self-test should be placed in the demo_drivers/mpi
   directory and where the output depends on the number of processors
   (e.g. for distributed problems) validata should be generated for
   two processors.

-- INTERFACE CHANGE (sort of): Provided the option to wipe the 
                    validata from the distribution. To make sure that
                    the self-test procedure doesn't break if there's
                    no validata, configure now assesses the
                    availability of validata by checking the existence
                    of demo_drivers/poisson/one_d_poisson/validata/one_d_poisson_results.dat.gz
                    If this is not found, we'll assume that there's
                    no validata anywhere. In that case, "make check" will still
                    compile and run the demo drivers but not use
                    fpdiff.py to compare them against the validata.
                    The behaviour is thus similar to what's done if
                    we don't have python. I therefore modified the 
                    "no_python" argument that used to be passed
                    to all validate.sh scripts to suppress the
                    execution of fpdiff.py and replaced it by
                    "no_fpdiff". As a result all validate.sh scripts
                    had to be changed. 

                      MAKE SURE YOU UPDATE ANY NEWS ONES BEFORE 
                      ADDING THEM TO THE DISTRIBUTION!

                    To reflect the fact that the fpdiff-ing can now
                    be suppressed for multiple reasons, renamed

                         bin/validate_no_python.sh
                    
                    to

                         bin/dummy_validate.sh

-- Clarified that various messages that appear on the screen
   during the build process are issued by oomph-lib.

-- Added additional safety checks to the optional deletion of 
   an existing build directory. For starters, the suggestion for the
   name of the non-default directory is now /home/joe_user/build rather than
   /home/joe_user. Also, if we're offering to wipe the existing
   build directory, the user now has to go through a windows-like
   "are you sure?" step (after being shown the contents of the
   directory) to make sure we don't accidentally delete the
   user's home directory (at his/her request, but still...).
   Thanks to Cedric for spotting this potential trap (the hard way;
   luckily he'd just made a backup...).



      




git-svn-id: svn+ssh://vummath.ma.man.ac.uk/home/svn-oomphlib/public/trunk@19 595e1d82-adf1-4282-bac7-b6073fa2ca8e

+++ autogen.sh
@@ -189,7 +189,7 @@ echo " "
 OptionPrompt " Is this OK? [y/n -- default: n]"
 reply=`OptionRead`
 if test "$reply" != "y" -a "$reply" != "Y" ; then 
-   OptionPrompt "Specify build directory [e.g. /home/joe_user] :"
+   OptionPrompt "Specify build directory [e.g. /home/joe_user/build] :"
    build_dir=`OptionRead`
 else
     echo "It's ok"
@@ -223,13 +223,28 @@ echo " "
 #---------------------
 if (test -d  $build_dir); then 
     echo " "
-    echo "Build directory exists"
-    OptionPrompt " Do you want to wipe it [y/n -- default: n]"
+    echo "Note: Build directory " $build_dir " exists."
+    echo " "
+    OptionPrompt "Do you want to wipe it [y/n -- default: n]"
     reply=`OptionRead`
     if test "$reply" = "y" -o "$reply" = "Y" ; then 
-       echo "Wiping it..."
-       rm -f -r $build_dir
-       echo "Done"
+
+       echo " "
+       echo "Sorry to be over-cautious here, but we'd better double-check"
+       echo "before we delete some of your precious files..."
+       echo " "
+       echo "The contents of " $build_dir " are:"
+       echo " "
+       ls -l  $build_dir
+       echo " "
+       OptionPrompt "Are you still sure you want to wipe it [y/n -- default: n]"
+       reply2=`OptionRead`
+       if test "$reply2" = "y" -o "$reply2" = "Y" ; then 
+          echo " "
+          echo "Wiping it..."
+          rm -f -r $build_dir
+          echo "Done"
+       fi
     fi
 fi
